backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.121.0
pytest==8.0.2
pytest-env==1.1.1
httpx>=0.24.0,<=0.26.0
python-jose[cryptography]==3.3.0
itsdangerous==2.1.2
pytest-mock>=3.10.0
pytest-httpx>=0.27.0
pytest-asyncio>=0.21.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.6
alembic==1.10.0
